Software PWM support for Fans and Dew Heaters

+++ indi_allsky/devices/fans/fanPwm.py
@@ -49,9 +49,9 @@ class FanPwm(FanBase):
 
 
         if not self.invert_output:
-            new_duty_cycle = ((2 ** 16) - 1) * new_state_i / 100
+            new_duty_cycle = int(((2 ** 16) - 1) * new_state_i / 100)
         else:
-            new_duty_cycle = ((2 ** 16) - 1) * (100 - new_state_i) / 100
+            new_duty_cycle = int(((2 ** 16) - 1) * (100 - new_state_i) / 100)
 
 
         logger.warning('Set fan state: %d%%', new_state_i)
